+++ clang/docs/ReleaseNotes.rst
@@ -532,6 +532,7 @@ clang-format
 ------------
 
 - Adds ``BreakBinaryOperations`` option.
+- Adds ``TemplateNames`` option.
 
 libclang
 --------